backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
motor==3.6.0
pymongo==4.9.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
pypdf==5.1.0
pdfminer.six
python-docx==1.1.2
passlib[bcrypt]==1.7.4
firebase-admin==6.5.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.150.0
openai==1.51.0
weasyprint==62.3
Pillow==11.0.0
jinja2==3.1.4
aiofiles==24.1.0
slowapi==0.1.9
groq
spacy
scikit-learn
